Move HostSettingsPage from component to common
