+++ assets/js/common/HostInfoBox/HostInfoBox.jsx
