+++ assets/js/common/HostInfoBox/HostInfoBox.test.jsx
