+++ assets/js/common/HostInfoBox/index.js
@@ -1,0 +1,3 @@
+import HostInfoBox from './HostInfoBox';
+
+export default HostInfoBox;
